For final results of the SUSY RPV B-L stop search, with AB2.4.29

+++ xAODAnaHelpers/JetSelector.h
@@ -136,7 +136,7 @@ private:
   int m_numEvent;         //!
   int m_numObject;        //!
   int m_numEventPass;     //!
-  int m_weightNumEventPass; //!
+  float m_weightNumEventPass; //!
   int m_numObjectPass;    //!
   int m_pvLocation;       //!
