+++ xAODAnaHelpers/MuonSelector.h
@@ -91,7 +91,7 @@ private:
   int m_numEvent;           //!
   int m_numObject;          //!
   int m_numEventPass;       //!
-  int m_weightNumEventPass; //!
+  float m_weightNumEventPass; //!
   int m_numObjectPass;      //!
 
   // cutflow
